+++ pkg/device-plugin/nvidiadevice/nvinternal/plugin/register.go
@@ -25,6 +25,7 @@ import (
 	"k8s.io/klog/v2"
 
 	"4pd.io/k8s-vgpu/pkg/api"
+	"4pd.io/k8s-vgpu/pkg/device/nvidia"
 	"4pd.io/k8s-vgpu/pkg/util"
 )
 
@@ -79,8 +80,8 @@ func (r *NvidiaDevicePlugin) RegistrInAnnotation() error {
 		return err
 	}
 	encodeddevices := util.EncodeNodeDevices(*devices)
-	annos[util.NodeHandshake] = "Reported " + time.Now().String()
-	annos[util.NodeNvidiaDeviceRegistered] = encodeddevices
+	annos[nvidia.HandshakeAnnos] = "Reported " + time.Now().String()
+	annos[nvidia.RegisterAnnos] = encodeddevices
 	klog.Infoln("Reporting devices", encodeddevices, "in", time.Now().String())
 	err = util.PatchNodeAnnotations(node, annos)