+++ pkg/util/nodelock/nodelock.go
@@ -1,55 +1,23 @@
-package util
+package nodelock
 
 import (
 	"context"
 	"fmt"
-	"os"
-	"path/filepath"
 	"time"
 
+	"4pd.io/k8s-vgpu/pkg/util/client"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/klog/v2"
 )
 
-var kubeClient kubernetes.Interface
-var DevicesToHandle []string
-
-func init() {
-	kubeClient, _ = NewClient()
-	DevicesToHandle = []string{}
-	DevicesToHandle = append(DevicesToHandle, NvidiaGPUCommonWord)
-	DevicesToHandle = append(DevicesToHandle, CambriconMLUCommonWord)
-}
-
-func GetClient() kubernetes.Interface {
-	return kubeClient
-}
-
-// NewClient connects to an API server
-func NewClient() (kubernetes.Interface, error) {
-	kubeConfig := os.Getenv("KUBECONFIG")
-	if kubeConfig == "" {
-		kubeConfig = filepath.Join(os.Getenv("HOME"), ".kube", "config")
-	}
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		klog.Infoln("InClusterConfig failed", err.Error())
-		config, err = clientcmd.BuildConfigFromFlags("", kubeConfig)
-		if err != nil {
-			klog.Errorln("BuildFromFlags failed", err.Error())
-			return nil, err
-		}
-	}
-	client, err := kubernetes.NewForConfig(config)
-	return client, err
-}
+const (
+	NodeLockTime = "4pd.io/mutex.lock"
+	MaxLockRetry = 5
+)
 
 func SetNodeLock(nodeName string) error {
 	ctx := context.Background()
-	node, err := kubeClient.CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
+	node, err := client.GetClient().CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
 	if err != nil {
 		return err
 	}
@@ -58,18 +26,18 @@ func SetNodeLock(nodeName string) error {
 	}
 	newNode := node.DeepCopy()
 	newNode.ObjectMeta.Annotations[NodeLockTime] = time.Now().Format(time.RFC3339)
-	_, err = kubeClient.CoreV1().Nodes().Update(ctx, newNode, metav1.UpdateOptions{})
+	_, err = client.GetClient().CoreV1().Nodes().Update(ctx, newNode, metav1.UpdateOptions{})
 	for i := 0; i < MaxLockRetry && err != nil; i++ {
 		klog.ErrorS(err, "Failed to update node", "node", nodeName, "retry", i)
 		time.Sleep(100 * time.Millisecond)
-		node, err = kubeClient.CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
+		node, err = client.GetClient().CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
 		if err != nil {
 			klog.ErrorS(err, "Failed to get node when retry to update", "node", nodeName)
 			continue
 		}
 		newNode := node.DeepCopy()
 		newNode.ObjectMeta.Annotations[NodeLockTime] = time.Now().Format(time.RFC3339)
-		_, err = kubeClient.CoreV1().Nodes().Update(ctx, newNode, metav1.UpdateOptions{})
+		_, err = client.GetClient().CoreV1().Nodes().Update(ctx, newNode, metav1.UpdateOptions{})
 	}
 	if err != nil {
 		return fmt.Errorf("setNodeLock exceeds retry count %d", MaxLockRetry)
@@ -80,7 +48,7 @@ func SetNodeLock(nodeName string) error {
 
 func ReleaseNodeLock(nodeName string) error {
 	ctx := context.Background()
-	node, err := kubeClient.CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
+	node, err := client.GetClient().CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
 	if err != nil {
 		return err
 	}
@@ -90,18 +58,18 @@ func ReleaseNodeLock(nodeName string) error {
 	}
 	newNode := node.DeepCopy()
 	delete(newNode.ObjectMeta.Annotations, NodeLockTime)
-	_, err = kubeClient.CoreV1().Nodes().Update(ctx, newNode, metav1.UpdateOptions{})
+	_, err = client.GetClient().CoreV1().Nodes().Update(ctx, newNode, metav1.UpdateOptions{})
 	for i := 0; i < MaxLockRetry && err != nil; i++ {
 		klog.ErrorS(err, "Failed to update node", "node", nodeName, "retry", i)
 		time.Sleep(100 * time.Millisecond)
-		node, err = kubeClient.CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
+		node, err = client.GetClient().CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
 		if err != nil {
 			klog.ErrorS(err, "Failed to get node when retry to update", "node", nodeName)
 			continue
 		}
 		newNode := node.DeepCopy()
 		delete(newNode.ObjectMeta.Annotations, NodeLockTime)
-		_, err = kubeClient.CoreV1().Nodes().Update(ctx, newNode, metav1.UpdateOptions{})
+		_, err = client.GetClient().CoreV1().Nodes().Update(ctx, newNode, metav1.UpdateOptions{})
 	}
 	if err != nil {
 		return fmt.Errorf("releaseNodeLock exceeds retry count %d", MaxLockRetry)
@@ -112,7 +80,7 @@ func ReleaseNodeLock(nodeName string) error {
 
 func LockNode(nodeName string) error {
 	ctx := context.Background()
-	node, err := kubeClient.CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
+	node, err := client.GetClient().CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
 	if err != nil {
 		return err
 	}